feat: rainbowkit in registration flow
