+++ src/authentication/create-wallet-account/styles.scss
@@ -13,13 +13,18 @@
   @include base();
 
   &__main {
-    box-sizing: border-box;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 24px;
     width: 100%;
-    padding: 8px;
+    padding: 24px;
+
+    box-sizing: border-box;
   }
 
   &__error {
-    padding: 4px 8px;
     font-size: 14px;
     line-height: 17px;
   }
