+++ src/authentication/web3-login/index.tsx
@@ -1,13 +1,12 @@
 import * as React from 'react';
 
 import { Alert } from '@zero-tech/zui/components';
-import { Button, Variant } from '@zero-tech/zui/components/Button/Button';
+import { Button } from '@zero-tech/zui/components/Button/Button';
 
 import { bemClassName } from '../../lib/bem';
 import './styles.scss';
 import { Web3LoginErrors } from '../../store/login';
-import { ConnectButton } from '@rainbow-me/rainbowkit';
-import { IconChevronRight } from '@zero-tech/zui/icons';
+import { RainbowKitConnectButton } from '../../lib/web3/rainbowkit/button';
 
 const cn = bemClassName('web3-login');
 
@@ -50,34 +49,3 @@ export class Web3Login extends React.Component<Web3LoginProperties, Web3LoginSta
     );
   }
 }
-
-interface RainbowKitConnectButtonProperties {
-  isDisabled?: boolean;
-}
-
-const RainbowKitConnectButton = ({ isDisabled }: RainbowKitConnectButtonProperties) => {
-  return (
-    <ConnectButton.Custom>
-      {({ account, openAccountModal, openConnectModal }) => {
-        if (account?.address) {
-          return (
-            <Button
-              isDisabled={isDisabled}
-              variant={Variant.Secondary}
-              onPress={openAccountModal}
-              endEnhancer={<IconChevronRight isFilled={true} />}
-            >
-              {account.displayName}
-            </Button>
-          );
-        } else {
-          return (
-            <Button isDisabled={isDisabled} onPress={openConnectModal}>
-              Connect Wallet
-            </Button>
-          );
-        }
-      }}
-    </ConnectButton.Custom>
-  );
-};
